requirements.txt
uvicorn
fastapi
pydantic
python-dotenv
groq
langchain_openai
langchain_groq
defusedxml
langchain_community
pypdf
pymilvus
langchain
langchain-google-genai
nltk
scikit-learn
numpy
python-docx
langchain-huggingface
sentence-transformers
langchain_chroma
duckduckgo-search
langgraph
langchain-tavily